+++ lib/features/branches/presentation/bloc/branches_event.dart
@@ -30,10 +30,6 @@ class GetRepositoryBranches extends BranchesEvent {
   GetRepositoryBranches();
 }
 
-class GetGraphCommits extends BranchesEvent {
-  GetGraphCommits();
-}
-
 class CreateNewBranchAndCheckout extends BranchesEvent {
   final String branchName;